Support ConfirmListeners in what I hope is a suitable vertx way.
Remove unnecessary logging from unit test.

Add a channel instance number to message confirmations

Implement reliable message publisher

Support for consumers reconnecting when the server disconnects unexpectedly

Give it a bit longer to do the initial connect to see whether that helps the CI builds

Change the connectionEstablishedCallbacks to be asynchronously trackable

Minor change to make publishers more efficient when there is a backlog of confirmations

Add asynchronous stop method to RabbitMQPublisher to enable clients to clear out the send queue and shut down nicely.

Fix RabbitMQPublisher.stop so it works when queue is empty.

Should not pass null in to a Handler<AsyncResult>

Some disconnect scenarios cause publish result handlers to fail, so catch any exceptions there in order that the queue not remain paused

Documentation for the Connection-Established callbacks

Adjust documentation for the Connection-Established callbacks

Reword the options around retrying for consistency

Remove RabbitMQConfirmListener and rename one basicPublish method to basicPublishWithDeliveryTag so it can have a sibling method that returns a Future

I found an error condition in the RabbitMQClientImpl that meant that consumers were unreliable when faced with updating clusters.
The node currently being used is taken down, the client reconnects correctly to another node that is being taken down - this generates an error that is not retried.
I've altered the logic here so that consumers retry with the same logic as reconnections.

I've also removed the failed call to handlers in publishers after a message failure.
This is because the RabbitMQPublisherImpl doesn't really have the concept of a failed message - it just holds it internally and retries until it succeeds.
If the handler passed in to the publisher is a Promise then it can only have a single call to fail() or succeed(), which isn't consistent with retrying.

Update version of rabbitmq-java-client to include fix for https://github.com/rabbitmq/rabbitmq-java-client/issues/525

Add check for retry configuration in RabbitMQClientImpl

Reacting to comments on PR

Modifications following review

Update to reflect upstream changes