I found an error condition in the RabbitMQClientImpl that meant that consumers were unreliable when faced with updating clusters. The node currently being used is taken down, the client reconnects correctly to another node that is being taken down - this generates an error that is not retried. I've altered the logic here so that consumers retry with the same logic as reconnections.
I've also removed the failed call to handlers in publishers after a message failure.
This is because the RabbitMQPublisherImpl doesn't really have the concept of a failed message - it just holds it internally and retries until it succeeds.
If the handler passed in to the publisher is a Promise then it can only have a single call to fail() or succeed(), which isn't consistent with retrying.